Create goats.php
Here is the main page. It should have buttons for all of the functions. It sends the input to goat1.php. The buttons on the bottom will go to other pages. I separated some of the functions out to make the layout look better it should also make the functions a little bit easier.